% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lv.kable.R
\name{lv.kable}
\alias{lv.kable}
\title{Extrahiert bei MC-Fragen das Label der Antwortoption}
\usage{
lv.kable(
  x,
  caption = NULL,
  col.width = "default",
  bold = TRUE,
  bold.col1 = TRUE,
  digits = 2,
  escape = FALSE,
  striped = TRUE
)
}
\arguments{
\item{x}{Ein dataframe.}

\item{col.width}{Vektor mit den Breiten der verschiedenen Spalten}

\item{bold}{Soll die Kopfzeile (welche die Spaltennamen enthält) header fett sein?}
}
\description{
Extrahiert bei MC-Fragen das Label der Antwortoption
}
\examples{
lv.kable(head(mtcars, 10))
}
